fix(auth): change employeeId type from Integer to Long in JwtAuthenticationFilter and JwtUtil

+++ Tick-It-User-Service/src/main/java/com/harish/TickIt/UserService/authentication/JwtAuthenticationFilter.java
@@ -41,7 +41,7 @@ public class JwtAuthenticationFilter extends org.springframework.web.filter.Once
 																						  .map(r-> new org.springframework.security.core.authority.SimpleGrantedAuthority(r))
 																						  .collect(java.util.stream.Collectors.toSet());
 					String username = jwtUtil.getUsernameFromToken(token);
-					Integer id= jwtUtil.getEmployeeId(token);
+					Long id= jwtUtil.getEmployeeId(token);
 					
 					UserPrincipal up= new UserPrincipal();
 					up.setEmployeeId(id);
